dissolve `scrapy.contrib_exp`
